requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.1
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.1
attrs==25.3.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==4.9.0
azure-identity==1.19.0
azure-search==1.0.0b2
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
distro==1.9.0
docutils==0.21.2
et_xmlfile==2.0.0
filelock==3.18.0
filetype==1.2.0
Flask==2.3.3
Flask-Cors==4.0.0
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.5.1
fuzzywuzzy==0.18.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
gunicorn==21.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.6
httpx==0.27.2
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Kivy==2.3.1
Kivy-Garden==0.1.5
Levenshtein==0.21.1
# Editable install with no version control (llm-safe==0.1.0)
-e /Users/vieirama/Desktop/llm_safe_snapshot_generator
llvmlite==0.44.0
lxml==5.3.2
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
mypy_extensions==1.1.0
narwhals==1.32.0
networkx==3.5
numba==0.61.2
numpy==2.2.4
oauthlib==3.2.2
openai==1.73.0
openai-whisper==20240930
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.15
opencensus-ext-flask==0.8.2
opencensus-ext-requests==0.8.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.241009
pathlib==1.0.1
pathspec==0.12.1
pdf2image==1.17.0
pillow==11.1.0
platformdirs==4.3.7
playwright==1.52.0
pluggy==1.5.0
portalocker==2.10.1
priority==2.0.0
propcache==0.2.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.9.0
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
PyPDF2==3.0.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-Levenshtein==0.21.1
pytz==2025.2
Quart==0.20.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
rumps==0.4.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.44.0
streamlit-feedback==0.1.3
sympy==1.14.0
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
torch==2.7.1
tornado==6.4.2
tqdm==4.67.0
types-pytz==2024.2.0.20241003
typing-inspection==0.4.1
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.17.1
